refactor: make PositionMixin properties sync, add Lit tests
