+++ packages/overlay/test/position-mixin-lit.test.js
@@ -1,0 +1,3 @@
+import './position-mixin-styles.js';
+import './position-mixin-lit.js';
+import './position-mixin.common.js';
