+++ packages/overlay/test/position-mixin-polymer.test.js
@@ -1,0 +1,3 @@
+import './position-mixin-styles.js';
+import './position-mixin-polymer.js';
+import './position-mixin.common.js';
