+++ packages/overlay/test/position-mixin-styles.js
@@ -1,0 +1,17 @@
+import { css } from 'lit';
+import { registerStyles } from '@vaadin/vaadin-themable-mixin/register-styles';
+
+registerStyles(
+  'vaadin-positioned-overlay',
+  css`
+    @keyframes slidein {
+      0% {
+        transform: translateY(10px);
+      }
+    }
+
+    :host(.animated) [part='overlay'] {
+      animation: slidein 0.2s;
+    }
+  `,
+);
